+++ CHANGELOG.md
@@ -19,6 +19,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - In some cases Wireframe API wasn't able to catch errors; this part should now work more consistently.
 - Enabled endpoints setting in Wireframe API module config now works as expected.
 
+## [0.18.3] - 2020-11-30
+
+### Fixed
+- Fixed an issue where Wireframe Tracy panel displayed errors if opened on the Admin.
+
 ## [0.18.2] - 2020-11-28
 
 ### Fixed